requirements.txt
absl_py==1.0.0
boto3==1.21.4
botocore==1.24.5
config==0.5.1
filelock==3.3.2
ftfy==6.1.1
h5py==2.10.0
hydra-core==1.1.0
hydra-colorlog==1.2.0
jieba==0.42.1
matplotlib==3.4.3
numpy==1.19.5
omegaconf==2.1.1
packaging==21.2
pandas==1.3.4
pythainlp==3.0.8
pytorch_lightning==1.5.1
pytorch_metric_learning==0.9.99
regex==2021.11.10
requests==2.26.0
rich==12.4.4
sacremoses==0.0.46
scikit_learn==1.1.1
scipy==1.7.2
sentencepiece==0.1.96
six==1.16.0
spacy==3.3.1
tokenizers==0.5.0
tqdm==4.62.3
uvicorn==0.18.1
wandb==0.12.6